requirements.txt
fastapi
uvicorn
python-binance
pandas
streamlit
requests 